Adapt to undocumented(?)/unexpected configuration behavior in Bitnami Tomcat.
It turns out that Bitnami Tomcat ignores “setenv.bat”, even though it creates
the file. Instead, it relies solely on tomcat8, the Windows-specific service
configuration and control application provided w/ Tomcat 8 on Windows.
(Observed with Bitnami Tomcat 8.5.24-2 on Windows 7.)

To adapt, for each fix-up made to "setenv.bat", a corresponding call is made
to tomcat8 for the Tomcat service on Windows.